requirements.txt
face-recognition==1.3.0
Pillow==10.0.1
numpy==1.24.3
psutil==5.9.5
Flask==2.3.3
Werkzeug==2.3.7
